Add VoiceActivityDetector base class
Add base class to allow different VAD implementations
